requirements.txt
addict==2.4.0
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-secret-manager==2.12.6
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.2
iniconfig==1.1.1
osrs==0.0.5
osrs-highscores==0.7.1
OSRS-Hiscores==0.5
OSRSBytes==1.3.0
packaging==21.3
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.11
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytz==2022.6
requests==2.27.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.6
